Add 'rsync-via-ssh'
rsync-via-ssh is a script I designed to sync files and directories between
to machines (e.g. a laptop and a desktop) using rsync over a SSH connection.

The directories which are synchronized are hardcoded inside the script's
'run_rsync' function, because that suits my use case.

Currently the script has one mandatory argument: the IP address of the
remote host (the destination of the sync). Optional arguments are:
the port on which the SSH server on the remote host is listening,
a custom filename for the log file genarated by the script.

An interactive mode is in development.