Add 'backup-to-HDD-1.sh'
backup-to-HDD-1.sh performs a backup of a set of specified directories
to an external hard drive.

The set of directories, and their specific options, are hardcoded in
the script's `sync_multimedia_dirs` and `sync_data_dirs` functions.

The destination of the backup (the mount point of the external drive)
can be specified as a parameter on the command line (defaults to
'/media/luca/HDD-1').